test: migrate some tests from the `driver` directory from Jest to AVA
