+++ packages/zwave-js/package.json
@@ -141,6 +141,7 @@
     "@types/jest": "^29.0.2",
     "@types/node": "^14.18.36",
     "@types/proper-lockfile": "^4.1.2",
+    "@types/proxyquire": "^1.3.28",
     "@types/semver": "^7.3.12",
     "@types/sinon": "^10.0.13",
     "@types/source-map-support": "^0.5.6",
@@ -154,6 +155,7 @@
     "jest-extended": "^2.0.0",
     "mockdate": "^3.0.5",
     "prettier": "^2.8.1",
+    "proxyquire": "^2.1.3",
     "sinon": "^14.0.0",
     "typescript": "4.9.4"
   }
