+++ packages/zwave-js/src/lib/driver/Statistics.test.ava.ts
@@ -1,4 +1,5 @@
 import { Mixin } from "@zwave-js/shared";
+import test from "ava";
 import EventEmitter from "events";
 import { StatisticsHost } from "./Statistics";
 
@@ -22,12 +23,10 @@ interface Test extends TestStatisticsHost {}
 @Mixin([TestStatisticsHost])
 class Test extends EventEmitter {}
 
-describe("lib/driver/Statistics", () => {
-	it("the statistics property is available and has the correct defaults", () => {
-		const test = new Test();
-		expect(test.statistics).toEqual({
-			one: 1,
-			two: 2,
-		});
+test("the statistics property is available and has the correct defaults", (t) => {
+	const test = new Test();
+	t.deepEqual(test.statistics, {
+		one: 1,
+		two: 2,
 	});
 });
